Add Cooperative Grid variant of GPU-STUMP
This variant uses a single kernel invocation to compute the matrix
profile and matrix profile indices. The loop over the range is pushed
into the cooperative kernel, and threads synchronize on iterations of
the loop using a grid sync.

The cooperative kernel can be faster than using multiple non-cooperative
launches, because the overhead of launching many individual kernels is
removed.

The choice of whether to use the cooperative kernel is left up to the
user, as not all GPUs / systems support cooperative launches.