Added both methods for upstream fetching, #414
Both methods, fetching locally and fetching through Github, are now
covered in Contribute.ipynb. The screenshot demonstrating the latter
method was changed to a screenshot of a stumpy fork. The instructions
for synching fork was also moved up to be between sections that
describe pushing changes and starting a pull requests.