Update top-k profile by getting insertion index
In NearestNeighbor case, the distance  between sequence i and its NN
is the smallest. However, due to `imprecision` in calculation, it is
possible that its corresponding distance, i.e. distance  between seq i
and its NN, is not the smallest value in its  top-k neighbors. So,
instead of inserting it at index 0, we use numpy.searchsorted to find
the correct insertion index.